refactor: don't use bundledDependencies
As described here:

    https://github.com/liferay/liferay-frontend-projects/pull/379

the use of "bundledDependencies" here is causing all sorts of pain when
trying to update Senna in a Yarn workspace (and pretty much anywhere
else), because it reduces our ability to influence what gets
installed/used. Given that we (Liferay) control the entire ecosystem, it
doesn't make much sense for us to tie our own hands in this way, so I'm
removing the "bundledDependencies".

In any case, these were probably redundant now, as explained here:

    https://stackoverflow.com/a/25044361/2103996

> When not to use bundled dependencies
>
> When developing, I don't think that the main point is to prevent
> accidental updates though. We have better tools for that, namely code
> repositories (git, mercurial, svn...) or now lock files.
>
> To pin your package versions, you can use:
>
> Option 1: Use the newer NPM version 5 that comes with node 8. It uses
> a package-lock.json file (see the node blog and the node 8 release)

(Note that we are already using a package-lock.json.)