test: bump NodeJS version to 10.15.1
To match what we are currently using in most liferay-portal branches. I
don't know if this will fix the CI failures:

-   https://travis-ci.org/github/metal/gulp-metal/builds/767729372
-   https://travis-ci.org/github/metal/gulp-metal/builds/767729343

but it makes sense to move onto a newer base before trouble-shooting
further; not that v10 is all that new nowadays...

    https://nodejs.org/en/about/releases/

notes that v10 is reaching End-Of-Life real soon now (2021-04-30).